chore: update input header file

+++ include/input.h
@@ -1,7 +1,7 @@
 #ifndef INPUT_H
 #define INPUT_H
 
-int getcmd(char** cmd);
+ssize_t getcmd(char** cmd);
 void reset(int* argc, char** cmd, char*** argv);
 
 #endif
